+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Only accretive/fixative changes will be made from now on.
 
 The following changes have been committed to the **master** branch since the 1.0.1 release:
 
+* Fix #40 by adding `next.jdbc.prepare/execute-batch!`.
 * Improved docstrings and documentation, especially around prepared statement handling.
 
 ## Stable Builds